+++ player.h
@@ -7,8 +7,8 @@ class Player: public Entity{
         int btnUp, btnDown;
         int velocity;
     public:
-    Player(int x, int y, int w, int h, unsigned int c, int btn1, int btn2, int vel);
-    Player(int x, int y, int w, int h, unsigned int c, int s);
+    Player(float x, float y, int w, int h, unsigned int c, int btn1, int btn2, int vel);
+    Player(float x, float y, int w, int h, unsigned int c, int s);
     Player();
     void setScore(int s);
     int getScore();
